// src/enum/prTemplate/prTemplates.enum.ts
export enum PrType {
  CAPEX = 'CAPEX',
  OPEX = 'OPEX',
}

export enum ViewerType {
  VIEWER = 'viewer',
  ASSOCIATE_OWNER = 'associate owner',
}

export enum PrTemplateStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum PrListSearchEnum {
  PR_ID = 'id',
  PR_TYPE = 'prType',
  PR_STATUS = 'prStatus',
  CREATED_AT = 'createdAt',
}

export enum PurchaseOrgEnum {
  UMWT = 'UMWT',
  ASSB = 'ASSB',
  AISB = 'AISB',
  TBU = 'TBU',
}

export enum DeliveryType {
  LIST = 'LIST',
  MANUAL = 'MANUAL',
}